Crystal-GFN refactoring, improvements and Stack environment

+++ scripts/crystal/eval_crystalgflownet.py
@@ -1,7 +1,6 @@
 """
 Computes evaluation metrics and plots from a pre-trained GFlowNet model.
 """
-
 import pickle
 import shutil
 import sys
@@ -14,8 +13,9 @@ from tqdm import tqdm
 
 sys.path.append(str(Path(__file__).resolve().parent.parent))
 
+from crystalrandom import generate_random_crystals
 from gflownet.gflownet import GFlowNetAgent
-from gflownet.utils.common import load_gflow_net_from_run_path, read_hydra_config
+from gflownet.utils.common import load_gflow_net_from_run_path
 from gflownet.utils.policy import parse_policy_config
 
 
@@ -23,25 +23,14 @@ def add_args(parser):
     """
     Adds command-line arguments to parser
 
-    Returns
-    -------
-    argparse.ArgumentParser
-        The parser with added arguments
+    Returns:
+        argparse.Namespace: the parsed arguments
     """
     parser.add_argument(
         "--run_path",
         default=None,
         type=str,
         help="Path to Hydra run containing config.yaml",
-    )
-    parser.add_argument(
-        "--conditional_env_config_path",
-        default=None,
-        type=str,
-        help="Path to a configuration YAML file containing the properties of an "
-        "environment to be used to do conditional sampling, that is constrain the "
-        "action space at sampling time. If the file is stored in the same directory "
-        "as the main config, the argument may be just the file name (not a path).",
     )
     parser.add_argument(
         "--n_samples",
@@ -79,33 +68,25 @@ def add_args(parser):
         action="store_true",
         help="Sample from an untrained GFlowNet",
     )
+    parser.add_argument(
+        "--random_crystals",
+        action="store_true",
+        help="Sample crystals uniformly, without constraints",
+    )
     parser.add_argument("--device", default="cpu", type=str)
     return parser
 
 
 def get_batch_sizes(total, b=1):
     """
-    Batches an iterable into chunks of size n and returns their expected lengths.
+    Batches an iterable into chunks of size n and returns their expected lengths
 
-    Example
-    -------
+    Args:
+        total (int): total samples to produce
+        b (int): the batch size
 
-    .. code-block:: python
-
-        >>> get_batch_sizes(10, 3)
-        [3, 3, 3, 1]
-
-    Parameters
-    ----------
-    total : int
-        total samples to produce
-    b : int
-        the batch size
-
-    Returns
-    -------
-    list
-        list of batch sizes
+    Returns:
+        list: list of batch sizes
     """
     n = total // b
     chunks = [b] * n
@@ -118,10 +99,8 @@ def print_args(args):
     """
     Prints the arguments
 
-    Parameters
-    ----------
-    args : argparse.Namespace
-        the parsed arguments
+    Args:
+        args (argparse.Namespace): the parsed arguments
     """
     print("Arguments:")
     darg = vars(args)
@@ -135,18 +114,6 @@ def set_device(device: str):
         return torch.device("cuda")
     else:
         return torch.device("cpu")
-
-
-def path_compatible(str):
-    """
-    Replace all non-alphanumeric characters with underscores
-
-    Parameters
-    ----------
-    str : str
-        The string to be made compatible
-    """
-    return "".join([c if c.isalnum() else "_" for c in str])
 
 
 def main(args):
@@ -168,61 +135,16 @@ def main(args):
 
     base_dir = Path(args.output_dir or args.run_path)
 
-    # ---------------------------------
-    # -----  Test GFlowNet model  -----
-    # ---------------------------------
-
-    if not args.samples_only:
-        gflownet.logger.test.n = args.n_samples
-        eval_results = gflownet.evaluator.eval()
-
-        # TODO-V: legacy -> ok to remove?
-        # keys = ["True reward and GFlowNet samples", "GFlowNet KDE Policy", "Reward KDE"]
-        # fignames = ["samples", "kde_gfn", "kde_reward"]
-
-        output_dir = base_dir / "figures"
-        print("output_dir: ", str(output_dir))
-        output_dir.mkdir(parents=True, exist_ok=True)
-
-        for figname, fig in eval_results["figs"].items():
-            output_fig = output_dir / (path_compatible(figname) + ".pdf")
-            if fig is not None:
-                fig.savefig(output_fig, bbox_inches="tight")
-        print(f"Saved figures to {output_dir}")
-
-        # Print metrics
-        print("Metrics:")
-        for k, v in eval_results["metrics"].items():
-            print(f"\t{k}: {v:.4f}")
-
     # ------------------------------------------
     # -----  Sample GFlowNet  -----
     # ------------------------------------------
 
-    # Read conditional environment config, if provided
-    # TODO: implement allow passing just name of config
-    if args.conditional_env_config_path is not None:
-        config_cond_env = read_hydra_config(
-            config_name=args.conditional_env_config_path
-        )
-        if "env" in config_cond_env:
-            config_cond_env = config_cond_env.env
-        env_cond = instantiate(
-            config_cond_env,
-            proxy=env.proxy,
-            device=config.device,
-            float_precision=config.float_precision,
-        )
-    else:
-        env_cond = None
-
-    # Handle output directory
     output_dir = base_dir / "eval" / "samples"
     output_dir.mkdir(parents=True, exist_ok=True)
     tmp_dir = output_dir / "tmp"
     tmp_dir.mkdir(parents=True, exist_ok=True)
 
-    if args.n_samples > 0 and args.n_samples <= 1e5:
+    if args.n_samples > 0 and args.n_samples <= 1e6:
         print(
             f"Sampling {args.n_samples} forward trajectories",
             f"from GFlowNet in batches of {args.sampling_batch_size}",
@@ -230,31 +152,59 @@ def main(args):
         for i, bs in enumerate(
             tqdm(get_batch_sizes(args.n_samples, args.sampling_batch_size))
         ):
-            batch, times = gflownet.sample_batch(
-                n_forward=bs, env_cond=env_cond, train=False
-            )
+            batch, times = gflownet.sample_batch(n_forward=bs, train=False)
             x_sampled = batch.get_terminating_states(proxy=True)
-            energies = env.proxy(x_sampled)
+            energies = env.oracle(x_sampled)
             x_sampled = batch.get_terminating_states()
-            df = pd.DataFrame(
-                {
-                    "readable": [env.state2readable(x) for x in x_sampled],
-                    "energies": energies.tolist(),
-                }
-            )
-            df.to_csv(tmp_dir / f"gfn_samples_{i}.csv")
-            dct = {"x": x_sampled, "energy": energies.tolist()}
+            dct = {"energy": energies.tolist()}
             pickle.dump(dct, open(tmp_dir / f"gfn_samples_{i}.pkl", "wb"))
 
         # Concatenate all samples
-        print("Concatenating sample CSVs")
-        df = pd.concat([pd.read_csv(f) for f in tqdm(list(tmp_dir.glob("*.csv")))])
-        df.to_csv(output_dir / f"{prefix}_samples.csv")
-        dct = {"x": [], "energy": []}
-        for f in tqdm(list(tmp_dir.glob("*.pkl"))):
+        print("Concatenating data")
+        dct = {"energy": []}
+        for f in tqdm(list(tmp_dir.glob("gfn_samples_*.pkl"))):
             tmp_dict = pickle.load(open(f, "rb"))
             dct = {k: v + tmp_dict[k] for k, v in dct.items()}
-        pickle.dump(dct, open(output_dir / f"{prefix}_samples.pkl", "wb"))
+        pickle.dump(dct, open(output_dir / f"{prefix}_energies.pkl", "wb"))
+
+        if "y" in input("Delete temporary files? (y/n)"):
+            shutil.rmtree(tmp_dir)
+
+    # ------------------------------------
+    # -----  Sample random crystals  -----
+    # ------------------------------------
+
+    # Sample random crystals uniformly without constraints
+    if args.random_crystals and args.n_samples > 0 and args.n_samples <= 1e6:
+        tmp_dir.mkdir(parents=True, exist_ok=True)
+        print(f"Sampling {args.n_samples} random crystals without constraints...")
+        for i, bs in enumerate(
+            tqdm(get_batch_sizes(args.n_samples, args.sampling_batch_size))
+        ):
+            x_sampled = generate_random_crystals(
+                n_samples=args.n_samples,
+                elements=config.env.composition_kwargs.elements,
+                min_elements=2,
+                max_elements=5,
+                max_atoms=config.env.composition_kwargs.max_atoms,
+                max_atom_i=config.env.composition_kwargs.max_atom_i,
+                space_groups=config.env.space_group_kwargs.space_groups_subset,
+                min_length=0.0,
+                max_length=1.0,
+                min_angle=0.0,
+                max_angle=1.0,
+            )
+            energies = env.oracle(env.states2proxy(x_sampled))
+            dct = {"energy": energies.tolist()}
+            pickle.dump(dct, open(tmp_dir / f"randomcrystals_samples_{i}.pkl", "wb"))
+
+        # Concatenate all samples
+        print("Concatenating data")
+        dct = {"energy": []}
+        for f in tqdm(list(tmp_dir.glob("randomcrystals_samples_*.pkl"))):
+            tmp_dict = pickle.load(open(f, "rb"))
+            dct = {k: v + tmp_dict[k] for k, v in dct.items()}
+        pickle.dump(dct, open(output_dir / f"randomcrystals_energies.pkl", "wb"))
 
         if "y" in input("Delete temporary files? (y/n)"):
             shutil.rmtree(tmp_dir)
